[#169, #173, #182][BACK][도서] - 도서 관계 테이블(출판사, 태그, 카테고리) CRUD 구현 및 카테고리 로직 변경 (#198)
* feat(도서): 출판사, 태그 CRUD http 테스트

* feat(도서): 도서 출판사 CRUD 구현

# Conflicts:
#	backend-server/src/main/java/shop/bluebooktle/backend/book/entity/BookPublisher.java
#	backend-server/src/main/java/shop/bluebooktle/backend/book/repository/BookPublisherRepository.java

* feat(도서): 도서 출판사 CRUD 구현

* fix(도서): 카테고리 수정

* fix(도서):출판사 삭제 로직 수정

# Conflicts:
#	backend-server/http/book.http

* feat(도서): 도서태그 관계테이블 CRUD

* feat(도서): 도서태그 관계테이블 CRUD

* fix(도서): Category Test 코드 수정

* fix(도서): Category Http Test 코드 수정 및 로직 수정, Entity 수정

* fix(도서): Category Http Test 코드 수정 및 로직 수정, Entity 수정

* feat(도서): book_category http파일 수정

* fix(도서): PR 코드 리뷰 보완(Transactional 어노테이션 추가)

+++ backend-server/src/main/java/shop/bluebooktle/backend/book/service/impl/PublisherServiceImpl.java
@@ -9,19 +9,22 @@ import lombok.RequiredArgsConstructor;
 import shop.bluebooktle.backend.book.dto.request.PublisherRequest;
 import shop.bluebooktle.backend.book.dto.response.PublisherInfoResponse;
 import shop.bluebooktle.backend.book.entity.Publisher;
+import shop.bluebooktle.backend.book.repository.BookPublisherRepository;
 import shop.bluebooktle.backend.book.repository.PublisherRepository;
 import shop.bluebooktle.backend.book.service.PublisherService;
 import shop.bluebooktle.common.exception.book.PublisherAlreadyExistsException;
+import shop.bluebooktle.common.exception.book.PublisherCannotDeleteException;
 import shop.bluebooktle.common.exception.book.PublisherNotFoundException;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class PublisherServiceImpl implements PublisherService {
 
 	private final PublisherRepository publisherRepository;
+	private final BookPublisherRepository bookPublisherRepository;
 
 	@Override
-	@Transactional
 	public void registerPublisher(PublisherRequest request) {
 		if (publisherRepository.existsByName(request.getName())) {
 			throw new PublisherAlreadyExistsException("출판사명 : " + request.getName());
@@ -33,11 +36,9 @@ public class PublisherServiceImpl implements PublisherService {
 	}
 
 	@Override
-	@Transactional
 	public void updatePublisher(Long publisherId, PublisherRequest request) {
 		Publisher publisher = publisherRepository.findById(publisherId)
 			.orElseThrow(() -> new PublisherNotFoundException(publisherId));
-
 		publisher.setName(request.getName());
 		publisherRepository.save(publisher);
 	}
@@ -59,11 +60,13 @@ public class PublisherServiceImpl implements PublisherService {
 	}
 
 	@Override
-	@Transactional
 	public void deletePublisher(Long publisherId) {
 		Publisher publisher = publisherRepository.findById(publisherId)
 			.orElseThrow(() -> new PublisherNotFoundException(publisherId));
-
+		// 출판사에 등록된 도서가 있을 경우 삭제 불가
+		if (bookPublisherRepository.existsByPublisher(publisher)) {
+			throw new PublisherCannotDeleteException();
+		}
 		publisherRepository.delete(publisher);
 	}
 }
